renamed update to insert, added ability to update DB instead of re-inserting

+++ cli.py
@@ -25,10 +25,14 @@ client = pymongo.MongoClient(host=conf['mongo_host'],
                              connect=False)
 
 
-@argh.arg('-q', nargs='+', help='Query arguments to search, example: [-q Aphelion_dist:3.1948214 Num_opps:38].')
-@argh.arg('-db', default=conf['db_name'], help='Name of the DB, where MPCORB data is stored.')
-@argh.arg('-col', default=conf['col_name'], help='Name of the collection in DB, where MPCORB is stored.')
-@argh.arg('-dat', default=conf['dat_file'], help='Name of DAT file to perform search.')
+@argh.arg('-q', nargs='+',
+          help='Query arguments to search, example: [-q Aphelion_dist:3.1948214 Num_opps:38].')
+@argh.arg('-db', default=conf['db_name'],
+          help='Name of the DB, where MPCORB data is stored.')
+@argh.arg('-col', default=conf['col_name'],
+          help='Name of the collection in DB, where MPCORB is stored.')
+@argh.arg('-dat', default=conf['dat_file'],
+          help='Name of DAT file to perform search.')
 @argh.arg('--pretty', help='Result will be formatted if argument passed.')
 @argh.arg('--file', help='Indicates that search must be performed in file')
 @argh.arg('--first', help='Indicates that search must be performed in file')
@@ -63,7 +67,8 @@ def find(pretty=False, file=False, first=False, **kwargs):
     col = kwargs.get('col')
     dat = kwargs.get('dat')
     if not kwargs.get('q'):
-        logger.error('Query cannot be empty and should be passed via [-q] parameter.')
+        logger.error(
+            'Query cannot be empty and should be passed via [-q] parameter.')
         exit(1)
     query = parse_query(kwargs['q'])
     query = convert(query)
@@ -84,9 +89,12 @@ def find(pretty=False, file=False, first=False, **kwargs):
     return result
 
 
-@argh.arg('-url', default=conf['mpcorb_url'], help='URL for downloading MPCORB file')
-@argh.arg('-path', default=os.curdir, help='Path, where file will be downloaded.')
-@argh.arg('-name', default=conf['mpcorb_file'], help='Custom name for file that will be downloaded.')
+@argh.arg('-url', default=conf['mpcorb_url'],
+          help='URL for downloading MPCORB file')
+@argh.arg('-path', default=os.curdir,
+          help='Path, where file will be downloaded.')
+@argh.arg('-name', default=conf['mpcorb_file'],
+          help='Custom name for file that will be downloaded.')
 @argh.arg('--extract', help='Extract downloaded file if it is archive.')
 def get(extract=False, **kwargs):
     """
@@ -109,14 +117,17 @@ def get(extract=False, **kwargs):
     """
     filename = kwargs.get('name')
     if not filename.endswith(('.gz', '.zip')) and extract:
-        logger.warning('Cannot determine type of compression, extraction can fail.')
+        logger.warning(
+            'Cannot determine type of compression, extraction can fail.')
     path = kwargs.get('path')
     if not os.path.exists(path):
-        logger.error("Failed to start download: no such directory: [{}]".format(path))
+        logger.error(
+            "Failed to start download: no such directory: [{}]".format(path))
         exit(1)
     url = kwargs.get('url')
     if not os.access(path, os.W_OK):
-        logger.error('Insufficient permission to write into folder [{}]'.format(path))
+        logger.error(
+            'Insufficient permission to write into folder [{}]'.format(path))
         exit(1)
     path = '{}/{}'.format(path, filename)
     logger.info('Downloading from {}'.format(url))
@@ -137,11 +148,14 @@ def get(extract=False, **kwargs):
 
 
 @argh.arg('-db', default=conf['db_name'], help='Name of the DB for updating.')
-@argh.arg('-col', default=conf['col_name'], help='Name of the collection in DB for updating.')
+@argh.arg('-col', default=conf['col_name'],
+          help='Name of the collection in DB for inserting.')
 @argh.arg('-path', help='Path to the file that will be used for DB update.')
-def update(**kwargs):
+@argh.arg('--update', help='Indicates that updating existing records must be '
+                           'performed instead of re-inserting full DB.')
+def insert(update=False, **kwargs):
     """
-    Function for updating database from JSON file.
+    Function for inserting data to database from JSON file.
     Drops existing collection records and replace it with new records.
     Used values from `config` by default.
     CLI args:
@@ -149,10 +163,12 @@ def update(**kwargs):
         -col name of the collection for data population
         -path: path to the data file
     Example:
-        `$ python cli.py update -path mpcorb_extended.json`
-        `$ python cli.py update -path mpcorb.json -col mpc_backup -db mpc2`
+        `$ python cli.py insert -path mpcorb_extended.json`
+        `$ python cli.py insert -path mpcorb.json -col mpc_backup -db mpc2`
+        `$ python cli.py insert -path mpcorb.json -col mpc_backup -db mpc2 --update`
 
     :param kwargs: keyword arguments from CLI.
+    :param update: indicates that existing records will be updated.
     """
     db_name = kwargs.get('db')
     col = kwargs.get('col')
@@ -169,25 +185,39 @@ def update(**kwargs):
         try:
             logger.info('Reading JSON file...')
             data = json_util.loads(f.read())
-            logger.info('DB: [{}] will be used.'.format(db_name))
-            if col in client[db_name].collection_names():
-                logger.warning('Existing collection [{}] will be dropped before update.'.format(col))
-            else:
-                logger.info('Creating new collection [{}]'.format(col))
+            logger.info('DB [{}] will be used.'.format(db_name))
+            if col in client[db_name].collection_names() and not update:
+                logger.warning(
+                    'Existing collection [{}] will be dropped before update.'.format(
+                        col))
             collection = client[db_name][col]
             if not data:
                 raise JSONDecodeError('', '', 1)
-            collection.drop()
-            logger.info('Inserting MPCORB data records into database, wait...')
-            for record in data:
-                collection.insert_one(record)
-            logger.info('Inserted {} records into [{}.{}]'.format(collection.count(),
-                                                                  db_name, col))
+            if update:
+                logger.info(
+                    'Updating MPCORB data records into database, wait...')
+                for record in data:
+                    collection.update_one(
+                        {'Principal_desig': record['Principal_desig']},
+                        {'$set': record}, upsert=False)
+                count = len(data)
+            else:
+                collection.drop()
+                logger.info(
+                    'Inserting MPCORB data records into database, wait...')
+                for record in data:
+                    collection.insert_one(record)
+                count = collection.count()
+            logger.info(
+                'Processed {} records into [{}.{}]'.format(count,
+                                                           db_name, col))
         except (JSONDecodeError, UnicodeDecodeError):
-            logger.error('Failed to read JSON file: data is empty or corrupted.')
+            logger.error(
+                'Failed to read JSON file: data is empty or corrupted.')
 
 
 if __name__ == '__main__':
     parser = argh.ArghParser()
-    parser.add_commands([find, get, update])
+    parser.add_commands([find, get, insert])
     parser.dispatch()
+
